refactor(app): decouple deposit pages
